mainProject/requirements.txt
Flask
google-generativeai
gTTS
speechrecognition
pydub

torch
opencv-python
numpy
pandas
Pillow
flask-cors